HACK HACK HACK HACK
This needs massive cleanup, but it's a snapshot of what I have right now, in
case people want to look at it.

I can build this on my Mac against stock Qt 4.8 libraries and a locally built
libmarble. I had to create tons of links for include directories and copy the
libmarble .h files around in order to be able to build - I'm sure there are
many things that aren't the way they should be.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>