Initial working version androidhelper.c
On android, Qt libraries are not to be used. But there are some important
functions implemented in qthelper.cpp which will be needed for subsurface
library on android. Hence, as a replacement, androidhelper.c will provide all
those functions. Exif related functions are not implemented yet as it would need
cpp. Instead, I think libexif should be used.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>